Time: 175 ms (59.43%), Space: 74.4 MB (42.35%) - LeetHub

+++ 1894-find-the-student-that-will-replace-the-chalk/1894-find-the-student-that-will-replace-the-chalk.cpp
@@ -1,10 +1,10 @@
 class Solution {
 public:
     int chalkReplacer(vector<int>& chalk, int k) {
-        long long int sum=0;
+        long sum=0;
         for(auto c:chalk) sum+=c;
-        long long int extra=k-(k/sum)*sum;
-        long long int index=0;
+        int extra=k-(k/sum)*sum;
+        int index=0;
         while(extra>=0){
             //cout<<index<<endl;
             if(extra>=chalk[index]){
